requirements.txt
# MCP core
mcp

# LangChain + MCP adapters
langchain
langchain-openai
langchain-mcp-adapters

# LangGraph (for tool orchestration)
langgraph

# UI
streamlit

# Environment variable support
python-dotenv

# OpenAI client (used by langchain-openai)
openai
